Share LATERAL joins between Postgres and MySQL
